requirements.txt
blinker==1.9.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
Flask==3.1.0
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.2
lxml==5.3.0
MarkupSafe==2.1.3
numpy==2.1.3
pandas==2.2.3
pillow==11.0.0
pip==22.2.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
requests==2.28.1
setuptools==63.2.0
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.13
Werkzeug==3.1.3
